+++ include/engine.h
@@ -4,6 +4,9 @@
 #include <camera.h>
 #include <window.h>
 
+// Struct to store the state of the engine
+// This is just so I can make the current scene, camera, and window global
+// So I can access them from systems to loop over the current scene
 struct EngineState
 {
     Scene scene;
@@ -21,4 +24,5 @@ struct EngineState
     }
 };  
 
+// Global state of the engine
 inline EngineState engineState;