+++ include/renderer.h
@@ -4,12 +4,23 @@
 #include <ecs.h>
 #include <shader.h>
 
+// 3D renderer for the engine
 namespace Renderer
 {
 
+// Intializes OpenGL (only makes the defaultShader at the moment)
 void Init();
+
+// Takes an entityID, gets its Transform and MeshRenderer components
+// and uses the data to render it to the screen
 void RenderModel(EntityID ent, const glm::mat4& projection, const glm::mat4& view);
 
+// Renders a line from one vec3 to another vec3, uses the line shader
+void RenderLine(glm::vec3 inPoint, glm::vec3 outPoint);
+
+// Default 3D shader
 inline Shader defaultShader;
+// Line shader, used for drawing 3d lines
+inline Shader lineShader;
 
 }